Modified observable to receive more arguments
The observable subroutine needs to receive the NN potential model and
the corresponding model parameters as arguments.

Some procedures in read_write had to be commented in order for the code
to compile

Also wrote functions to benchmark analytic derivatives (via chain
rule) against numerical approximations. All derivatives match within
the error of the numerical approximation. Functions are commented and
left for reference.

Signed-off-by: Rodrigo Navarro Perez <rnavarroperez@sdsu.edu>