Modified AV18
Made charge dependence consistent with the original paper. This reduced
the number of independent parameters from 44 to 42. This is still larger
than the 40 parameters mentioned in the paper. Asked Bob Wiringa for
clarification.

Added scattering lengths and deuteron to database

Added a print of parameters and chi^2 at every iteration of the
Levenberg Marqardt method.

Signed-off-by: Rodrigo Navarro Perez <rnavarroperez@sdsu.edu>