Fixed NaN for derivatives of data with low Energy
When the T_lab is very small (like in one of the np data points)
the derivatives of certain phasefhifts would sometimes would result in
NaN.

This was to the fact that when taking the derivative of a ratio (usually
of two very small numbers) taking the squared of the denominator would
become zero and result in the NaN or infinity. This was fixed by
changing the order of the operations, instead of squaring and dividing
we now divide twice by the denominator.

This allowed us to put that low energy data point back into the
data base.

For now we exclude the deuteron from the fits

Signed-off-by: Rodrigo Navarro Perez <rnavarroperez@sdsu.edu>