NaN check in observable and new stopping criteria
Similar changes to phaseshifts to avoids NaNs in the derivatives of
scattering observable.

Also added a check for for NaNs in the derivatives. If there's a NaN
in the array with the derivatives the input is printed to screen (
kinematics and parameters) and the program is stopped. This should help
troubleshoot if the same bug appears again.

Modified the stopping criteria in the optimization routine to
avoid being stuck between two almost identical points in parameter space
that result in different chi square. This is a result of the fact that
the systematic error contribution to the chi square may or may not be
added depending on its value

Signed-off-by: Rodrigo Navarro Perez <rnavarroperez@sdsu.edu>