Modified so that minimization is "faster"
The code excludes all data after 1965. This makes the calculation of
each chi square faster.

The stopping criteria in the chi square minization has been relaxed.
This allows the code to find a "minimum" faster.

The purpose of this changes is that explorations to the implementation
of the bootstrap can be made without having to wait a long time
for the calculation of each chi square and the optmization.

Signed-off-by: Rodrigo Navarro Perez <rnavarroperez@sdsu.edu>