[BugFix] Occasional NaNs in derivatives
In very rare occasions some parameters would result in NaN derivatives
for scattering observables. The root cause turned out to be in the
calculation of the coupled phase-shifts in the eigen representation

Sometimes a denominator would be exactly equal to zero. The formula
for the derivatives was analytically simplified for those cases and
implemented.

Signed-off-by: Rodrigo Navarro Perez <rnavarroperez@sdsu.edu>