Use unit in display functions. New output namelist
The functions that display the model's parameters now take a unit as
an argument. That unit can either be the `output_unit` constant from
`iso_fortran_env` for standard output to screen, or the unit of an
already opened file.

New `output` namelist to indicate whether or not to save results and
the name of the output file

Signed-off-by: Rodrigo Navarro Perez <rnavarroperez@sdsu.edu>