It is possible to adjust to phaseshifts now.
Phases can now be treated as 'observables'. Their type is the
name of the partial wave (i.e. 1s0, 3p0, 3s1, ep1, 3d1, etc).

For now there are 2 databases (files). One with all the real
experimental observables and with the average phaseshifts from the
collection of 'High Quality' potential and their standard deviation
as the uncertainty

Signed-off-by: "Rodrigo Navarro Perez" <"rnavarroperez@sdsu.edu">