Created default mask for av18
Right now not all parameters are used and therefore some derivatives
are zero. That makes the inversion of the alpha matrix fail (if no
observable depends on a parameter, there's no way to optimize such
parameter). To avoid this we use the mask so that the parameters that
are not used so far are not optimized.

Soon those dependencies will be added