Avoiding Lambda -> infinity
In some cases the parameter lambda in the optimization routine can
become so large that the new tried parameters are numerically the same
as the current best fit parameters. When that happens the code ends up
trying the same parameters over and over and even though the code is
already pretty close to a minimum.

To avoid this situation the condition to exit the optimization loop now
also allows to exit if lambda becomes large.